# Conflicts:
#	delete_user.php
#	list_users.php
#	models/UserModel.php

+++ models/UserModel.php
@@ -13,8 +13,9 @@ class UserModel extends BaseModel {
         return $user;
     }
 
-    public function findUser($keyword) {
-        $sql = 'SELECT * FROM users WHERE user_name LIKE %'.$keyword.'%'. ' OR user_email LIKE %'.$keyword.'%';
+    public function findUser($keyword)
+    {
+        $sql = 'SELECT * FROM users WHERE user_name LIKE %' . $keyword . '%' . ' OR user_email LIKE %' . $keyword . '%';
         $user = $this->select($sql);
 
         return $user;
@@ -26,23 +27,41 @@ class UserModel extends BaseModel {
      * @param $password
      * @return array
      */
-    public function auth($userName, $password) {
+    public function auth($userName, $password)
+    {
         $md5Password = md5($password);
-        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "'.$md5Password.'"';
+        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "' . $md5Password . '"';
 
         $user = $this->select($sql);
         return $user;
     }
+
+    // public function getUseById($id)
+    // {
+    //     # code...
+    //     $sql = 'SELECT `name` FROM users WHERE id = ' . $id;
+    //     $user = $this->select($sql);
+    //     return $user;
+    // }
 
     /**
      * Delete user by id
      * @param $id
      * @return mixed
      */
-    public function deleteUserById($id) {
-        $sql = 'DELETE FROM users WHERE id = '.$id;
-        return $this->delete($sql);
-
+    public function deleteUserById($id)
+    {
+        $user = $this->getUsers();
+        
+        foreach ($user as $value) {
+            if (($value['id']).md5("monweb1") == $id) {
+                // var_dump($id);
+                // var_dump($value['id'].md5("monweb1"));
+                // die();
+                $sql = 'DELETE FROM users WHERE id = ' . $value['id'];
+                return $this->delete($sql);
+            }
+        }
     }
 
     /**
@@ -50,7 +69,9 @@ class UserModel extends BaseModel {
      * @param $input
      * @return mixed
      */
-    public function updateUser($input) {
+
+    public function updateUser($input)
+    {
         $sql = 'UPDATE users SET 
                  name = "' . mysqli_real_escape_string(self::$_connection, $input['name']) . '"
                 ,`fullname`="' . mysqli_real_escape_string(self::$_connection, $input['fullname'])  . '"
@@ -81,10 +102,11 @@ class UserModel extends BaseModel {
      * @param array $params
      * @return array
      */
-    public function getUsers($params = []) {
+    public function getUsers($params = [])
+    {
         //Keyword
         if (!empty($params['keyword'])) {
-            $sql = 'SELECT * FROM users WHERE name LIKE "%' . $params['keyword'] .'%"';
+            $sql = 'SELECT * FROM users WHERE name LIKE "%' . $params['keyword'] . '%"';
 
             //Keep this line to use Sql Injection
             //Don't change
@@ -104,4 +126,3 @@ class UserModel extends BaseModel {
         self::$_instance = new self();
         return self::$_instance;
     }
-}
